contrib/jenkins.sh: update for aper-prefix tree
Add the missing 'autoreconf -fi', and disable 'make check' as checks on
this branch are failing with:
  fatal: making test-suite.log: failed to create ../tests/00-empty-OK.asn1.trs

I spent some time looking into it and don't see why it is doing that,
and it doesn't seem worth investing more time right now as we are not
actively developing on asn1c currently (and if we were it probably made
more sense to try to rebase on the upstream version / try to upstream
our additions?) so just disable 'make check'.

To explain what "this branch" is:
* We used to have a master branch (now old-master), which was based on
  2010 upstream code of asn1c with aper related patches on top. It had
  the contrib/jenkins.sh script and 'make check' was passing.
* In practice we've been using the "aper-prefix" branch for the past
  years (to generate files for osmo-iuh.git), which was based on 2015
  code from asn1c with the aper related patches and a patch to add
  ASN1C_PREFIX to avoid clashing names of generated files on top. There,
  no contrib/jenkins.sh file was present and 'make check' was failing.

So this new branch is the same as "aper-prefix", but with the
contrib/jenkins.sh file and now this patch to make it pass.

Related: OS#2435
Related: upstream development: https://github.com/vlm/asn1c
Change-Id: I79a398cba75339c308f56c2e44ea99b3776be164